- moves timeline classes into its own package
